Uploading test files
The pdf test file is for spring 2016. I tried  converting the pdf using PyPDF2, PdfFileReader and pdfminer with no luck. In my opinion pdfminer is the best out of them all. 
So I ended up using http://www.zamzar.com/convert/pdf-to-csv/
What is left to parse the data out is either moving some fields in the CSV file to align in their right column or have special rules in the parser.